+++ plugin/lib.ts
@@ -1,0 +1,5 @@
+
+export async function importModule(url: string, name: string) {
+	const module = await import(url);
+	return module[name];
+}